About to modify the header (side bar) rendering and position
