+++ app/src/main/java/com/example/culinar/GroceryActivity.kt
@@ -12,7 +12,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -22,23 +21,11 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.CutCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.AddCircle
-import androidx.compose.material.icons.filled.CheckCircle
-import androidx.compose.material.icons.filled.Clear
-import androidx.compose.material.icons.filled.DateRange
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Edit
-import androidx.compose.material.icons.filled.Email
-import androidx.compose.material.icons.filled.Face
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material.icons.filled.Person
+import androidx.compose.material.icons.filled.Info
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.filled.Settings
-import androidx.compose.material.icons.filled.ShoppingCart
-import androidx.compose.material.icons.outlined.ExitToApp
-import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.Icon
@@ -54,15 +41,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.viewmodel.viewModelFactory
 import com.example.culinar.ui.theme.CulinarTheme
 
 class GroceryActivity : ComponentActivity() {
@@ -71,7 +55,7 @@ class GroceryActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             CulinarTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                Scaffold(modifier = Modifier.fillMaxSize().background(color = Color(0xFFFFFFFF))) { innerPadding ->
                     Grocery(
                         modifier = Modifier.padding(innerPadding)
                     )
@@ -86,283 +70,14 @@ fun Grocery(modifier: Modifier = Modifier) {
     var screenOn by remember { mutableIntStateOf(1) }
 
     if (screenOn == 1) {
-        HomeTemplate(subscreen = { value -> GroceryList(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value })
+        HomeTemplate(subscreen = { value -> GroceryList(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value }, modifier = modifier)
     } else if (screenOn == 2) {
-        HomeTemplate(subscreen = { value -> GroceryAddItem(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value })
+        HomeTemplate(subscreen = { value -> GroceryAddItem(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value }, modifier = modifier)
     } else if (screenOn == 3) {
-        HomeTemplate(subscreen = { value -> GroceryModifyItem(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value })
+        HomeTemplate(subscreen = { value -> GroceryModifyItem(changeOnboardingScreen = value) }, changeOnboardingScreen = { value -> screenOn = value }, modifier = modifier)
     }
 }
 
-
-@Composable
-fun Header(modifier: Modifier = Modifier) {
-
-    var menuClicked by remember { mutableStateOf(false) }
-
-    Column () {
-        Spacer(modifier = Modifier.height(20.dp).background(color = Color(0xFF3CB460)))
-
-
-        Row(
-            modifier = modifier
-                .height(60.dp)
-                .fillMaxWidth()
-                .background(color = Color(0xFF3CB460)),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-
-            // Menu button
-            Button(
-                onClick = {menuClicked = !menuClicked},
-                modifier = Modifier.width(90.dp).height(90.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = if (!menuClicked) Color(0xFF3CB460) else Color(0xFF1C6231),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.Menu,
-                    contentDescription = "Menu",
-                    tint = Color.White,
-                    modifier = Modifier.width(40.dp).height(40.dp)
-                )
-
-            }
-
-            Spacer(modifier = Modifier.weight(1f))
-
-            /*
-            Text(
-                text = "Esteban GOMEZ",
-                fontSize = 20.sp,
-                fontFamily = FontFamily.Serif,
-                textAlign = TextAlign.Center,
-                fontWeight = FontWeight.Bold,
-                lineHeight = 21.sp,
-                modifier = Modifier.width(120.dp),
-            )
-            */
-            //Spacer(modifier = Modifier.weight(1f))
-            // Profile picture
-            Button(
-                onClick = {},
-                modifier = Modifier.width(90.dp).height(90.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF3CB460),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.Face,
-                    contentDescription = "Clear",
-                    tint = Color.White,
-                    modifier = Modifier.width(45.dp).height(45.dp)
-                )
-            }
-        }
-
-        // Side bar
-        if (menuClicked) {
-            Column(
-                modifier = Modifier
-                    .background(color = Color(0xFFFFFCF0))
-                    .width(150.dp)
-                    .fillMaxHeight(),
-                verticalArrangement = Arrangement.SpaceEvenly,
-
-            ) {
-                // Settings button
-                Column (
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    Button(
-                        onClick = { menuClicked = !menuClicked },
-                        modifier = Modifier
-                            .width(150.dp)
-                            .height(150.dp),
-                        shape = CutCornerShape(3.dp),
-                        colors = ButtonColors(
-                            containerColor = Color(0x003CB460),
-                            contentColor = Color.White,
-                            disabledContainerColor = Color(0x003CB460),
-                            disabledContentColor = Color.White
-                        )
-                    ) {
-                        Icon(
-                            Icons.Outlined.Settings,
-                            contentDescription = "Settings",
-                            tint = Color(0xFF0A0A0A),
-                            modifier = Modifier
-                                .width(70.dp)
-                                .height(70.dp)
-                        )
-                    }
-                    Text("Settings", fontSize = 20.sp)
-                }
-
-
-                // Log out button
-                Column (
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    Button(
-                        onClick = { menuClicked = !menuClicked },
-                        modifier = Modifier
-                            .width(150.dp)
-                            .height(150.dp),
-                        shape = CutCornerShape(3.dp),
-                        colors = ButtonColors(
-                            containerColor = Color(0x003CB460),
-                            contentColor = Color.White,
-                            disabledContainerColor = Color(0x003CB460),
-                            disabledContentColor = Color.White
-                        )
-                    ) {
-                        Icon(
-                            Icons.Outlined.ExitToApp,
-                            contentDescription = "Log out",
-                            tint = Color(0xFF0A0A0A),
-                            modifier = Modifier
-                                .width(70.dp)
-                                .height(70.dp)
-                        )
-                    }
-                    Text("Log out", fontSize = 20.sp)
-                }
-
-            }
-        }
-    }
-}
-
-@Composable
-fun Footer(modifier: Modifier = Modifier) {
-
-    Column {
-        Spacer(modifier = Modifier.weight(1f))
-
-        // Footer bar
-        Row(
-            modifier = modifier
-                .height(50.dp)
-                .fillMaxWidth()
-                .background(color = Color(0xFF3CB460)),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-
-            // Calendar button
-            Button(
-                onClick = {},
-                modifier = Modifier.width(72.dp).height(72.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF1C6231),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.DateRange,
-                    contentDescription = "Menu",
-                    tint = Color.White,
-                    modifier = Modifier.width(40.dp).height(40.dp)
-                )
-
-            }
-
-            // Groceries button
-            Button(
-                onClick = {},
-                modifier = Modifier.width(72.dp).height(72.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF3CB460),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.ShoppingCart,
-                    contentDescription = "Clear",
-                    tint = Color.White,
-                    modifier = Modifier.width(45.dp).height(45.dp)
-                )
-            }
-
-            // Recipes button
-            Button(
-                onClick = {},
-                modifier = Modifier.width(72.dp).height(72.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF3CB460),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.CheckCircle,
-                    contentDescription = "Clear",
-                    tint = Color.White,
-                    modifier = Modifier.width(45.dp).height(45.dp)
-                )
-            }
-
-            // Home button
-            Button(
-                onClick = {},
-                modifier = Modifier.width(72.dp).height(72.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF3CB460),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.Home,
-                    contentDescription = "Clear",
-                    tint = Color.White,
-                    modifier = Modifier.width(45.dp).height(45.dp)
-                )
-            }
-
-            // Community button
-            Button(
-                onClick = {},
-                modifier = Modifier.width(72.dp).height(72.dp).padding(0.dp),
-                shape = CutCornerShape(3.dp),
-                colors = ButtonColors(
-                    containerColor = Color(0xFF3CB460),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFF3CB460),
-                    disabledContentColor = Color.White
-                )
-            ) {
-                Icon(
-                    Icons.Default.Email,
-                    contentDescription = "Clear",
-                    tint = Color.White,
-                    modifier = Modifier.width(45.dp).height(45.dp)
-                )
-            }
-
-            //
-
-        }
-    }
-}
 
 
 @Composable
@@ -370,8 +85,8 @@ fun HomeTemplate(modifier: Modifier = Modifier, subscreen: @Composable ((Int) ->
     Box()
     {
         subscreen(changeOnboardingScreen)
-        Header()
-        Footer()
+        Header(modifier = modifier)
+        Footer(modifier = modifier)
     }
 }
 
@@ -379,12 +94,12 @@ fun HomeTemplate(modifier: Modifier = Modifier, subscreen: @Composable ((Int) ->
 fun GroceryList(modifier: Modifier = Modifier, changeOnboardingScreen: (Int) -> Unit) {
 
     // Retrieve list of grocery items
-    val grorecyItems = List(10) { index -> "item ${index.toString()}" }
+    val grorecyItems = List(10) { index -> "item ${index + 1}" }
 
     // Screen content
-    Column (verticalArrangement = Arrangement.SpaceEvenly, horizontalAlignment = Alignment.CenterHorizontally, modifier = Modifier.fillMaxSize()) {
+    Column (verticalArrangement = Arrangement.SpaceEvenly, horizontalAlignment = Alignment.CenterHorizontally, modifier = modifier.fillMaxSize()) {
         // Header place
-        Spacer(modifier = Modifier.height(80.dp))
+        Spacer(modifier = modifier.height(90.dp))
 
         // Body place
         // Title of the subscreen
@@ -397,26 +112,27 @@ fun GroceryList(modifier: Modifier = Modifier, changeOnboardingScreen: (Int) -> 
             lineHeight = 50.sp,
             color = Color(0xFF3CB460),
 
-            modifier = Modifier.fillMaxWidth()
+            modifier = modifier.fillMaxWidth()
                 .height(50.dp)
                 .background(color = Color(0xFFE5E3E3))
             )
 
-        Spacer(modifier = Modifier.height(15.dp))
+        Spacer(modifier = modifier.height(15.dp))
 
         // Search bar
+        var searchText by remember { mutableStateOf("") }
         TextField(
-            value = "",
+            value = searchText,
             placeholder = { Text("Rechercher un aliment") },
-            onValueChange = {},
+            onValueChange = {value -> searchText = value},
             leadingIcon = {Icon(Icons.Default.Search, contentDescription = "Rechercher")},
-            modifier = Modifier.width(300.dp).height(50.dp),
+            modifier = modifier.width(300.dp).height(50.dp),
         )
-        Spacer(modifier = Modifier.height(5.dp))
+        Spacer(modifier = modifier.height(5.dp))
         // GroceryItem add button
         Button(
             onClick = { changeOnboardingScreen(2) },
-            modifier = Modifier.width(160.dp).height(55.dp),
+            modifier = modifier.width(170.dp).height(55.dp),
             shape = CutCornerShape(3.dp),
             colors = ButtonColors(
                 containerColor = Color(0xFF59EA85),
@@ -426,14 +142,14 @@ fun GroceryList(modifier: Modifier = Modifier, changeOnboardingScreen: (Int) -> 
             )
         ) {
             Icon(Icons.Default.AddCircle, contentDescription = "Ajouter", tint = Color.White)
-            Spacer(modifier = Modifier.width(10.dp))
+            Spacer(modifier = modifier.width(10.dp))
             Text(text = "Ajouter", fontSize = 18.sp, letterSpacing = 3.sp)
         }
 
-        Spacer(modifier = Modifier.height(10.dp))
+        Spacer(modifier = modifier.height(10.dp))
         // Grocery items list
-        Column () {
-            LazyColumn (modifier = Modifier.fillMaxSize()) {
+        Column (modifier = modifier.weight(1f)) {
+            LazyColumn (modifier = modifier.fillMaxWidth()) {
                 items (grorecyItems) { item ->
                     GroceryItem(
                         item = item,
@@ -442,8 +158,10 @@ fun GroceryList(modifier: Modifier = Modifier, changeOnboardingScreen: (Int) -> 
                 }
             }
         }
-    }
 
+        // Footer space
+        Spacer(modifier = modifier.height(90.dp))
+    }
 }
 
 
@@ -452,15 +170,20 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
 
     Row (
         verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.padding(vertical = 2.dp, horizontal = 6.dp)
+        modifier = modifier.padding(vertical = 2.dp, horizontal = 6.dp)
             .fillMaxWidth()
             .height(108.dp)
             .border(color = Color(0xFFAAAAAA), width = 1.dp, shape = CutCornerShape(3.dp))
     ) {
         // Image of the item
-        Image(Icons.Default.Face, contentDescription = "Aliment", modifier = Modifier.width(85.dp).height(85.dp))
+        Image(Icons.Default.Info,
+            contentDescription = "Aliment",
+            modifier = modifier
+                .width(85.dp)
+                .height(85.dp)
+        )
 
-        Spacer(modifier = Modifier.weight(1f))
+        Spacer(modifier = modifier.weight(1f))
         // Details on the item
         Column {
             // Name of the item
@@ -478,7 +201,7 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
                     fontFamily = FontFamily.Serif,
                 )
 
-                Spacer(modifier = Modifier.width(40.dp))
+                Spacer(modifier = modifier.width(40.dp))
                 // Quantity of the item
                 Text(
                     text = "x 1",
@@ -488,11 +211,11 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
             }
 
         }
-        Spacer(modifier = Modifier.weight(1f))
+        Spacer(modifier = modifier.weight(1f))
 
         // Modify and delete buttons
         Column (
-            modifier = Modifier
+            modifier = modifier
                 .padding(vertical = 0.dp, horizontal = 5.dp)
                 .height(100.dp)
                 //.border(color = Color(0xFFAAAAAA), width = 1.dp, shape = CutCornerShape(3.dp))
@@ -502,7 +225,7 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
             // Delete button
             Button(
                 onClick = {},
-                modifier = Modifier.width(65.dp).height(48.dp).padding(vertical = 0.dp),
+                modifier = modifier.width(65.dp).height(48.dp).padding(vertical = 0.dp),
                 shape = CutCornerShape(3.dp),
                 colors = ButtonColors(
                     containerColor = Color(0xFFE91E63),
@@ -515,16 +238,16 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
                     Icons.Default.Delete,
                     contentDescription = "Supprimer",
                     tint = Color.White,
-                    modifier = Modifier.height(45.dp).width(45.dp)
+                    modifier = modifier.height(45.dp).width(45.dp)
                 )
             }
 
-            Spacer(modifier = Modifier.height(10.dp))
+            Spacer(modifier = modifier.height(10.dp))
 
             // Modify button
             Button(
                 onClick = { changeOnboardingScreen(3) },
-                modifier = Modifier.width(65.dp).height(35.dp)
+                modifier = modifier.width(65.dp).height(35.dp)
                     .border(color = Color(0xFF3F51B5), width = 2.dp, shape = CutCornerShape(3.dp)),
                 // shape = CutCornerShape(3.dp),
                 colors = ButtonColors(
@@ -536,7 +259,6 @@ fun GroceryItem(modifier: Modifier = Modifier, item : String, changeOnboardingSc
 
             ) {
                 Icon(Icons.Default.Edit, contentDescription = "Modifier", tint = Color(0xFF2196F3))
-
             }
         }
     }
